update: Integrated CPP contents in header file
Because Arduino IDE ignores .cpp files in other
directories it had to be integrated in the header file.
