+++ Code/ML/tinyML.h
@@ -16,9 +16,26 @@
 #define ARENA_SIZE 5000
 Eloquent::TF::Sequential<TF_NUM_OPS, ARENA_SIZE> tf;
 
-void ml_init(void);
+void ml_init(void) {
+  Serial.println("__TINYML_INIT__");
 
-float ml_predict(float *input);
-unsigned int ml_predict_latency(void);
+  /* Initialize model loader and check for errors. */
+  while (!tf.begin(model).isOk()) {
+    Serial.println(tf.exception.toString());
+  }
+}
+
+float ml_predict(float *input) {
+  /* Predict and check for errors */
+  if (!tf.predict(input).isOk()) {
+    Serial.println(tf.exception.toString());
+  }
+
+  return tf.output(0);
+}
+
+unsigned int ml_predict_time(void) {
+    return tf.benchmark.microseconds();
+}
 
 #endif
